requirements.txt
pillow
pandas
pyinstaller
